Added threading to API calls
Created a CollabApiThread class, which stores the results from the api call (call_gccollab_stats()) as an attribute, then pass those attributes on to the rest of the script.

API calls are now run "simultaneously," rather than sequentially, allowing the script to execute in less than half the normal time.